requirements.txt
fastapi
uvicorn[standard]
streamlit
pandas
openpyxl
langchain
langchain-community
langchain-google-genai
faiss-cpu
python-multipart
python-dotenv
numpy<2.0
bytez
